minor fix for submission list page
