+++ .gitignore
@@ -119,3 +119,4 @@ meetings/approvals/migrations/**
 meetings/conferences/migrations/**
 meetings/submissions/migrations/**
 meetings/api/migrations/**
+meetings/files/migrations/**
