+++ app/templates/conference/index.hbs
@@ -38,7 +38,7 @@
                             </div>
 
                             <!-- This is the basic code for a table of submissions without images -->
-                            {{#if model.submissionCount}}
+                            {{#if model.conf.submissionCount}}
                                 <table class="table table-hover">
                                     <thead class="thead">
                                         <tr>
