+++ meetings/osf_oauth2_adapter/provider.py
@@ -1,4 +1,4 @@
-from .apps import OsfOauth2AdapterConfig
+from meetings.utils import OsfOauth2AdapterConfig
 from allauth.socialaccount import providers
 from allauth.socialaccount.providers.base import ProviderAccount
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider
